requirements.txt
tensorflow==1.7
scipy==1.0.0
scikit-learn
opencv-python
h5py
matplotlib
Pillow
requests
psutil
apscheduler
